Refactor: ExpenseItem을 포함하는 ExpenseItemWithCategory, ExpenseItemWithDetails 클래스 생성

+++ data/expense/src/main/java/com/kappzzang/jeongsan/entity/expenselist/ExpenseRemoteEntity.kt
@@ -10,7 +10,7 @@ data class ExpenseRemoteEntity(
     @SerializedName("title")
     val title: String,
     @SerializedName("totalPrice")
-    val totalPrice: Long,
+    val totalPrice: Int,
     @SerializedName("createdAt")
     val createdAt: String,
     @SerializedName("state")
